Faster version of fix.
Note that this FAILS 3 tests, but only with respect to the error
messages being nicely formatted. Ordinarily I wouldn't commit without
either making it pass the tests, or at least changing the tests. But
this is a hot fix.  The message format in the tests is still
desirable, I just don't have time to figure out how to get it
efficiently. Meanwhile I have a user for whom downstream packages are
broken.